game over button, later improving the style
TODO: add a button to restart the buttons.

improve the style of the buttons its unnecessary to do all of this, add a better way of rendering them

Outside of that they seem to be doing good the only problem is that i need to add more classes to manage them more easily, so i dont have to waste too much time in this bs xd.

outside of that everything seems to be doing good or at least it seems to be the case.

thats all i guess.